requirements.txt
feedparser
openai-whisper
storage3
deeplake
streamlit
langchain
anthropic
jq